moved unused routes from Labs18 cohort into a separate folder, tested routes and notated non-working routes, S3 information required for two remaining routes
