+++ Lab18_unusedcode/ResetPasswordRouter.js
@@ -2,7 +2,7 @@
 const express = require("express");
 const router = express.Router();
 
-const Volunteer = require("../models/volunteer-model.js");
+const Volunteer = require("./volunteer-model.js");
 
 // To find the req.body of id and email to make sure that the email exists in the database
 router.get("/getvolunteer", (req, res) => {
