+++ api/authRouter.js
@@ -1,7 +1,7 @@
 const router = require("express").Router();
 const bcrypt = require("bcryptjs");
 
-const Volunteer = require("../models/volunteer-model.js");
+const Volunteer = require("../Lab18_unusedcode/volunteer-model.js");
 
 const generateToken = require("../middleware/Token.js");
 // const mw = require("../auth/auth-middleware.js");
